Add files via upload
Online Doctor Appointment System using PHP and Mysql is developed with point of manage the list of data of doctors, patients, appointment bookings, doctor available schedule information etc. This PHP Project on Doctor Appointment System, which is web based or online application which will resolves the problem of booking of Appointment according to the choice and requirement of the patient. Under this system, there is one feature like it is very effective solution which has offered with different available doctor availability on suitable dates and times, so patient can book doctor appointment according to their requirements or choice and patient can also cancel doctor appointment which also available under this system.